refactor(MovieDetailModal): 컴포넌트 클래스로 변경 및 이벤트 로직 수정
